requirements.txt
flask
fpdf
python-dotenv
google-generativeai
